feat(core): move cursor class enum to core improve line auto complete reaction
