+++ .changeset/curvy-walls-talk.md
@@ -1,0 +1,9 @@
+---
+'@plait/common': minor
+'@plait/core': minor
+'@plait/draw': minor
+'@plait/mind': minor
+---
+
+move cursor class enum to core
+improve line auto complete reaction
