+++ packages/common/src/constants/resize.ts
@@ -1,10 +1,3 @@
-export enum ResizeCursorClass {
-    ew = 'ew-resize',
-    ns = 'ns-resize',
-    nesw = 'nesw-resize',
-    nwse = 'nwse-resize'
-}
-
 export enum ResizeHandle {
     ne = 'ne',
     n = 'n',
