+++ packages/core/src/constants/cursor.ts
@@ -1,0 +1,10 @@
+export enum ResizeCursorClass {
+    ew = 'ew-resize',
+    ns = 'ns-resize',
+    nesw = 'nesw-resize',
+    nwse = 'nwse-resize'
+}
+
+export enum CursorClass {
+    crosshair = 'crosshair'
+}
