--- packages/core/src/constants/resize.ts
@@ -1,3 +1,0 @@
-export enum ResizeCursorClass {
-    'ew-resize' = 'ew-resize'
-}
