+++ packages/draw/src/plugins/with-line-auto-complete-reaction.ts
@@ -1,4 +1,4 @@
-import { PlaitBoard, PlaitElement, RgbaToHEX, drawCircle, isSelectionMoving, toPoint, transformPoint } from '@plait/core';
+import { CursorClass, PlaitBoard, PlaitElement, RgbaToHEX, drawCircle, isSelectionMoving, toPoint, transformPoint } from '@plait/core';
 import { PlaitDrawElement } from '../interfaces';
 import { getAutoCompletePoints, getHitIndexOfAutoCompletePoint, getSelectedDrawElements } from '../utils';
 import { GeometryComponent } from '../geometry.component';
@@ -7,18 +7,13 @@ import { LINE_AUTO_COMPLETE_HOVERED_DIAMETER, LINE_AUTO_COMPLETE_HOVERED_OPACITY
 
 export const withLineAutoCompleteReaction = (board: PlaitBoard) => {
     const { pointerMove } = board;
-
     let reactionG: SVGGElement | null = null;
-
     board.pointerMove = (event: PointerEvent) => {
         reactionG?.remove();
-        if (isSelectionMoving(board)) {
-            pointerMove(event);
-            return;
-        }
+        PlaitBoard.getBoardContainer(board).classList.remove(CursorClass.crosshair);
         const selectedElements = getSelectedDrawElements(board);
         const movingPoint = transformPoint(board, toPoint(event.x, event.y, PlaitBoard.getHost(board)));
-        if (!PlaitBoard.isReadonly(board) && selectedElements.length === 1 && PlaitDrawElement.isGeometry(selectedElements[0])) {
+        if (!PlaitBoard.isReadonly(board) && !isSelectionMoving(board) && selectedElements.length === 1 && PlaitDrawElement.isGeometry(selectedElements[0])) {
             const points = getAutoCompletePoints(selectedElements[0]);
             const hitIndex = getHitIndexOfAutoCompletePoint(movingPoint, points);
             const hitPoint = points[hitIndex];
@@ -32,6 +27,7 @@ export const withLineAutoCompleteReaction = (board: PlaitBoard) => {
                     fillStyle: 'solid'
                 });
                 PlaitBoard.getElementActiveHost(board).append(reactionG);
+                PlaitBoard.getBoardContainer(board).classList.add(CursorClass.crosshair);
             }
         }
         pointerMove(event);
